+++ src/index.ts
@@ -5,6 +5,7 @@ export * from "./types/props/BorderProps";
 export * from "./types/props/ColorProps";
 export * from "./types/props/LayoutProps";
 export * from "./types/props/TextProps";
+export * from "./types/props/IconProps";
 
 export * from "./types/tokens/buttonTokens";
 export * from "./types/tokens/linkTokens";
@@ -12,6 +13,7 @@ export * from "./types/tokens/formTokens";
 export * from "./types/tokens/typographyTokens";
 export * from "./types/tokens/safeAreaTokens";
 export * from "./types/tokens/footerTokens";
+export * from "./types/tokens/notificationTokens";
 
 export * from "./types/pipelinesolucoes";
 export * from "./utils/mergeThemeParts";