+++ src/types/props/IconProps.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export interface IconProps {
+  icon?: React.ReactNode;
+  color?: string;
+  size?: number | string;
+}